nhanesdata: Harmonized Access to NHANES Survey Data

Instant access to harmonized National Health and Nutrition Examination Survey (NHANES) data spanning 1999-2023. Retrieve pre-processed datasets from reliable cloud storage with automatic type reconciliation and integrated search tools for variables and datasets. Simplifies NHANES data workflows by handling cycle management and maintaining data consistency across survey waves. Data is sourced from https://www.cdc.gov/nchs/nhanes/.

See also

Author

Maintainer: Kyle Grealis kylegrealis@proton.me (ORCID)

Other contributors:

Amrit Baral abaral3@jhu.edu (ORCID) [contributor]

Natalie Neugaard nataliegoulett@gmail.com (ORCID) [contributor]

Raymond Balise balise@miami.edu (ORCID) [contributor]

Johannes Thrul jthrul@jhu.edu (ORCID) [contributor]

Janardan Devkota jdevkot1@jhu.edu (ORCID) [contributor]